Modifying LANE REVERSAL on PHB2 for both processors
Per Ric Mata, I needed to enable the lane reversal attribute on PHB2 for both processors to fix a bug in the processor.  This was not seen in the lab as it is typically a performance impact on the GPU management link.  This is a much bigger impact if the link were attached to an IO card.