Adding in HPC Bulk Power Limit Cap
Added missing HPC bulk power limit cap attribute.  Previously it was failing through to the N+1 power cap which was fine as it was set arbitrarily high.  Now this needs to be here to prevent the N+1 power cap from causing performance degradation on worst case config HPC (non-redundant systems).